Add files via upload
This the first step to anomaly detection and correction process. It finds the segments that are candidates for correction and summarizes the relevant data in preparation for the next step. This is also the most time consuming part of the process. In my experience, this code took about 24 hours to run for all 154 tilegroups on spark emr with approximately 125 concurrent processes.